requirements.txt
yfinance
numpy
pandas
streamlit
datetime
psycopg2
psycopg2-binary
pylint
pytest
python-dotenv